feat: use uint32_t instead of size_t

+++ include/pal-api.h
@@ -38,8 +38,8 @@
  * \retval          PAL_RC_INVALID_PARAM if `tx_capacity` or `max_msg_size` are invalid.
  */
 enum PalReturnCode pal_api_init(struct PalHandler *hpal,
-                                size_t rx_capacity,
-                                size_t max_msg_size,
+                                uint32_t rx_capacity,
+                                uint32_t max_msg_size,
                                 pal_deserialize_fn serialize,
                                 void (*cs_enter)(void),
                                 void (*cs_exit)(void),
@@ -58,7 +58,7 @@ enum PalReturnCode pal_api_init(struct PalHandler *hpal,
  * \retval          PAL_RC_QUEUE_FULL if the reception queue is full;
  * \retval          PAL_RC_IO_ERR if the "push-back" operation fails.
  */
-enum PalReturnCode pal_api_add_to_rx_queue(struct PalHandler *hpal, uint8_t *buff, size_t size);
+enum PalReturnCode pal_api_add_to_rx_queue(struct PalHandler *hpal, uint8_t *buff, uint32_t size);
 
 /*!
  * \brief           Pop the first message and exec the deserialization.
