+++ src/pal-api.c
@@ -33,8 +33,8 @@ EAGLETRT_STATIC_INLINE enum PalReturnCode prv_pal_deserialize_dummy(const struct
 }
 
 enum PalReturnCode pal_api_init(struct PalHandler *hpal,
-                                size_t rx_capacity,
-                                size_t max_msg_size,
+                                uint32_t rx_capacity,
+                                uint32_t max_msg_size,
                                 pal_deserialize_fn deserialize,
                                 void (*cs_enter)(void),
                                 void (*cs_exit)(void),
@@ -46,14 +46,14 @@ enum PalReturnCode pal_api_init(struct PalHandler *hpal,
         return PAL_RC_INVALID_PARAM;
     }
     hpal->deserialize = deserialize == NULL ? prv_pal_deserialize_dummy : deserialize;
-    ring_buffer_api_init(&hpal->rx_queue, max_msg_size + sizeof(size_t), rx_capacity, cs_enter, cs_exit, arena);
+    ring_buffer_api_init(&hpal->rx_queue, max_msg_size + sizeof(uint32_t), rx_capacity, cs_enter, cs_exit, arena);
     hpal->max_msg_size = max_msg_size;
-    hpal->add_to_rx_msg = (uint8_t *)arena_allocator_api_alloc(arena, max_msg_size + sizeof(size_t));
-    hpal->exec_rx_msg = (uint8_t *)arena_allocator_api_alloc(arena, max_msg_size + sizeof(size_t));
+    hpal->add_to_rx_msg = (uint8_t *)arena_allocator_api_alloc(arena, max_msg_size + sizeof(uint32_t));
+    hpal->exec_rx_msg = (uint8_t *)arena_allocator_api_alloc(arena, max_msg_size + sizeof(uint32_t));
     return PAL_RC_OK;
 }
 
-enum PalReturnCode pal_api_add_to_rx_queue(struct PalHandler *hpal, uint8_t *buff, size_t size) {
+enum PalReturnCode pal_api_add_to_rx_queue(struct PalHandler *hpal, uint8_t *buff, uint32_t size) {
     if (hpal == NULL || buff == NULL)
         return PAL_RC_NULL_PTR;
 
@@ -66,8 +66,8 @@ enum PalReturnCode pal_api_add_to_rx_queue(struct PalHandler *hpal, uint8_t *buf
     if (size > hpal->max_msg_size)
         return PAL_RC_MSG_TOO_BIG;
 
-    memcpy(hpal->add_to_rx_msg, &size, sizeof(size_t));
-    memcpy((uint8_t *)hpal->add_to_rx_msg + sizeof(size_t), buff, size);
+    memcpy(hpal->add_to_rx_msg, &size, sizeof(uint32_t));
+    memcpy((uint8_t *)hpal->add_to_rx_msg + sizeof(uint32_t), buff, size);
     RingBufferReturnCode res = ring_buffer_api_push_back(&hpal->rx_queue, hpal->add_to_rx_msg);
     if (res == RING_BUFFER_FULL) {
         return PAL_RC_QUEUE_FULL;
